5_Hackaton/requirements.txt
label-studio
pillow
streamlit
ultralytics